requirements.txt
pandas
numpy
matplotlib
scikit-learn
nltk
spacy
jupyter
notebook
requests
biopython
wordcloud
plotly